+++ test/runtests.jl
@@ -14,6 +14,8 @@ include("entropy_dimension.jl")
 include("nlts_tests.jl")
 # Periodicity:
 include("periodicity_tests.jl")
+# Chaos Detection:
+include("chaos_detection_tests.jl")
 
 ti = time() - ti
 println("\nTest took total time of:")